@@ -1,24 +1,20 @@
-# type: ignore
 # Configuration file for the Sphinx documentation builder.
 #
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
+# For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
-# -- Path setup --------------------------------------------------------------
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#
-import os
+import pathlib
 import sys
 
-sys.path.insert(0, os.path.abspath("../../new_modeling_toolkit"))
+from sphinxawesome_theme.postprocess import Icons
 
+print(pathlib.Path(__file__).parents[2] / "tests")
+sys.path.insert(0, str(pathlib.Path(__file__).parents[2]))
 
 # -- Project information -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = "Resolve"
-copyright = "2023, Energy & Environmental Economics, Inc."
+project = "Kit"
+copyright = "2024, Energy & Environmental Economics, Inc."
 author = "Energy & Environmental Economics, Inc."
 
 
@@ -29,92 +25,122 @@ author = "Energy & Environmental Economics, Inc."
 # ones.
 extensions = [
     "myst_parser",
+    "sphinxawesome_theme.highlighting",
+    "sphinxcontrib.mermaid",
     "sphinx.ext.autodoc",
     "sphinx.ext.graphviz",
+    "sphinx.ext.intersphinx",
     "sphinx.ext.mathjax",
     "sphinx.ext.napoleon",
     "sphinx.ext.todo",
-    "sphinx_copybutton",
     "sphinxcontrib.autodoc_pydantic",
     "sphinx_design",
+    "sphinx_last_updated_by_git",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ["../_templates"]
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
 # exclude_patterns = []
 
-# autodoc_member_order = "bysource"
-# autodoc_pydantic_model_member_order = "bysource"
+# Intentionally not inheriting all autodoc members, because I want to be able to control what pydantic model
+# attributes are shown (e.g., not attrs inherited from`Component` or `CustomModel`)
+autodoc_default_options = {
+    "inherited-members": False,
+}
 
-# autodoc_pydantic_model_show_json = False
-# autodoc_pydantic_model_show_config_summary = False
-# autodoc_pydantic_model_show_config_member = False
+autodoc_pydantic_field_list_validators = False
+autodoc_pydantic_model_show_config_member = False
+autodoc_pydantic_model_show_config_summary = False
+autodoc_pydantic_model_show_json = False
+autodoc_pydantic_model_show_validator_members = False
+autodoc_pydantic_model_show_validator_summary = False
+autodoc_pydantic_model_signature_prefix = "class"
+
+add_module_names = False
+toc_object_entries_show_parents = "hide"
+
+intersphinx_mapping = {"pandas": ("https://pandas.pydata.org/pandas-docs/stable/", None)}
+myst_url_schemes = ["https"]
 
 # -- Options for HTML output -------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = "furo"
-html_favicon = "_images/e3-logo.ico"
+html_theme = "sphinxawesome_theme"
+html_favicon = "e3-logo.ico"
+
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
-html_css_files = ["css/furo-e3.css"]
+html_css_files = ["css/e3.css"]
 
-html_title = "Resolve"
+html_title = "Kit"
+html_permalinks_icon = Icons.permalinks_icon
 
-
-# Dark mode colors need to be declared as theme "variables" here, which then can be called in furo-e3.css
-# https://github.com/pradyunsg/furo/discussions/485#discussioncomment-3209866
-html_theme_options = {
-    "light_css_variables": {
-        "color-brand-primary": "#034E6E",  # The standard E3 dark blue
-        "color-brand-background": "#E2ECF0",  # The E3 light blue/grey background color
-        "color-brand-dropdown": "white",
-    },
-    "light_logo": "resolve-logo-light.svg",
-    "dark_css_variables": {
-        "color-brand-primary": "#C4AD73",
-        "color-brand-background": "#212529",  # Same as sphinx-design --sd-color-dark
-        "color-brand-dropdown": "#212529",  # Same as sphinx-design --sd-color-dark
-    },
-    "dark_logo": "resolve-logo-dark.svg",
-    "sidebar_hide_name": True,
-    "top_of_page_button": "edit",
+html_sidebars = {
+    "branding": ["sidebar_main_nav_links.html"],
+    "dev": ["sidebar_main_nav_links.html"],
+    "roadmap": ["sidebar_main_nav_links.html"],
 }
+
+html_theme_options = {
+    "awesome_external_links": True,
+    "breadcrumbs_separator": Icons.chevron_right,
+    "logo_dark": "_static/logos/kit-dark.svg",
+    "logo_light": "_static/logos/kit-light.svg",
+    "main_nav_links": {
+        "Branding": "branding",
+        "Development Guide": "dev",
+        "Changelog & Roadmap": "roadmap",
+    },
+    "show_breadcrumbs": True,
+    "show_prev_next": True,
+    "show_scrolltop": True,
+}
+
+graphviz_output_format = "svg"
+
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 # html_static_path = ["_static"]
-
 add_module_names = False
 
-suppress_warnings = ["myst.header"]
+suppress_warnings = ["myst.header", "git.too_shallow"]
 
 
 myst_enable_extensions = [
     "amsmath",
-    # "attrs_inline",
-    "dollarmath",
+    "attrs_inline",
     "colon_fence",
+    "dollarmath",
+    "substitution",
 ]
+
+latex_elements = {"extrapackages": "\\usepackage{amsmath}"}
+
+# ------------ Mermaid configuration --------------------------------
+mermaid_d3_zoom = True
+
 
 # -- Instructions for how to update ------------------------------------------
 """
 1. If you need to update the autodoc files (e.g., structure changes),
    run the following command from the docs/ directory
-    sphinx-apidoc -f -o source [absolute path to src directory]
-2. Add text to the RST files manually for more documentation
-3. Run make html to update the HTML files
-4. Run make latex to update LaTeX files
+    sphinx-apidoc -f -o source [absolute path to new_modeling_toolkit directory]
+2. Convert the .rst files to .md using the command rst2myst 
+   (https://docs.readthedocs.io/en/stable/guides/migrate-rest-myst.html#how-to-convert-existing-restructuredtext-documentation-to-myst)
+2. Add text to the .md files manually for more documentation
+3. Run `make html` to update the HTML files
+4. Run `make latex` to update LaTeX files
 
 https://ethreesf.sharepoint.com/sites/Training/_layouts/OneNote.aspx?id=%2Fsites%2FTraining%2FSiteAssets%2FTraining%20Notebook&wd=target%28Technical%20Skills%20Training%2FPython.one%7C916C3A04-A4B3-4112-9E7F-F2F503E5B87C%2FDocumentation%20%26%20Docstrings%7C72764774-57EE-4BC6-8496-E63524712FF6%2F%29
 """
